QDoc: Remove ClangCodeParser::parseHeaderFile
When QDoc processes a project, it parses its code-base to extract and
sanitize the user provided documentation.

When a C++ project is processed, QDoc requires a list of headers files
that should be considered when parsing the code-base.

Currently, the list of headers files is provided by the user through the
project configuration.

Internally, QDoc uses its `Config` singleton, which is used to represent
the content of a configuration file, to obtain a list of headers that
should be considered.

`Config` is built as a generic key-value store, so that it provides
certain "raw" values that are later processed by its consumers.

The main loop for QDoc obtains this list, processes it slighlty and then
iterates it and passes each of the headers to
`ClangCodeParser::parseHeaderFile`.

`ClangCodeParser` is the component taking care of parsing C++ header and
source files.

`parseHeaderFile` does not in fact parse the header file. Indeed, QDoc
always uses a PCH file that it build in `ClangCodeParser::buildPCH` when
parsing the code-base for a C++ project.

Instead, `parseHeaderFile` performs some additional processing for the
list of headers and then store the results into a member variable of
`ClangCodeParser`.

To avoid having the processing of the header list split over multiple
components, to simplify the interface of `ClangCodeParser` and to
prepare for additional changes to `ClangCodeParser` internal state to
facilitate an effort to parallelize certain parts of QDoc,
`paseHeaderFile` is now removed.

The declaration of `parseHeaderFile` in `clangcodeparser.h` and its
definition in the relevant implementation file were removed.

A new member function of `Config`, `getHeaderFiles`, was added.
`getHeaderFiles` centralizes the processing we currently perform on the
list of header files, such as the removal of duplicates.

It returns a set of `HeaderFilePath`, a small structure that decomposes
the path to an header file, as extracted from the configuration file,
into its path and filename.
This decomposition is later used in `ClangCodeParser`, where access to
both the path and the filename of an header file is currently required.

The code taking performing the now centralized processing is removed.
The code extracting the list of header files and passing it to
`parseHeaderFile` in the main loop of QDoc is removed along with the single
usage of `parseHeaderFile` itself.

Previously, `ClangCodeParser::accepted_header_file_extension` was used
to filter the list of header files based on their extension.

As the processing is now centralized to `getHeaderFiles`,
`ClangCodeParser::accepted_header_file_extension` was removed and an
equivalent of it was introduced in the scope of `getHeaderFiles`, where
it is used for the same purpose.

The internal state that the various calls to `parseHeaderFile`
initialized is now initialized in the constructor for `ClangCodeParser`,
based on a usage of `getHeaderFiles` as called from the `Config`
dependency of the constructor.

Due to an implementation detail of `getHeaderFiles` which disallows it
from being a `const` member of `Config`, the dependency on `Config` for
the constructor of `ClangCodeParser` was modified to be a mutable
reference from a `const` reference.

`ClangCodeParser::m_allHeaders`, which contains the finalized list of
headers that was populated in `parseHeaderFile`, and is used all around
`ClangCodeParser`, was modified to respect the new type returned by
`getHeaderFiles`.

Usages of `m_allHeaders` were modified where required to work with the
new typing.

Task-number: QTBUG-111686
Change-Id: I3b554eaca246514d87d29a9082e5936b20df0c93
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>