QDoc: Move include paths processing outside ClangCodeParser
When QDoc processes a project, it parses its code-base to obtain and
sanitize the user-provided documentation for the project.

For projects that makes use of C++ code, QDoc uses Clang's APIs to
obtain the required information about the code-base.

Profiling shows that calls to Clang are the major bottleneck in QDoc at
the current time.

To reduce the time spent by Clang, QDoc is being moved to parallelize
and batch certain operations that are related to Clang calls.

`ClangCodeParser`, the component responsible for all of the calls to
Clang libraries, is currently architectured in a way that is not usable
for those optimization purposes.

Hence, it is slowly being modified to be better suited for this
purposes.
As part of those modification, we expect `ClangCodeParser` to be split
into multiple components.

Currently, `ClangCodeParser` is the entry point to a series of
operations that happen in different phases and have very different
restriction on their scopes and their dependencies.

Some of those operation shares some internal state that
an instance of `ClangCodeParser` keeps.

To allow for splitting those various operations up, we need to be able
to pass this shared data around the code-base to all relevant parties.

One such datum is `ClangCodeParser::m_includePaths`, which is at various
times to instrument the include paths for a compilation performed by
Clang's APIs.

`ClangCodeParser` performs some relatively expensive processing to
obtain those include paths.
Due to the current structure of the code, this processing is performed a
single time by `ClangCodeParser` itself, stored and used multiple times.

We don't want to perform the required processing more than required as
it can be expensive, such that the processing that `ClangCodeParser` was
performing is now moved to the main loop, where it can later be passed
around as required.

The processing steps for `m_includePaths` that were performed in the
constructor for `ClangCodeParser` is now moved to the main loop of a
QDoc execution.

The constructor for `ClangCodeParser` was modified to obtain a reference
to the required include paths data and store it in `m_includePaths`,
whose type was changed accordingly, instead of processing it itself.

The type of the data was slightly changed, from a `QList` to an
`std::vector` and usages of it were changed accordingly.

The "utilities.h" header is now not included by "clangcodeparser.cpp"
anymore.
It was included only to access certain operations required by the
processing of the include paths data and is now unnecessary.

Task-number: QTBUG-111686
Change-Id: Ia86bacf34dbefbc74ae908fa8c402ba2286627e4
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>