QDoc: Move defines processing outside ClangCodeParser
When QDoc processes a project, it parses its code-base to obtain and
sanitize the user-provided documentation for the project.

For projects that makes use of C++ code, QDoc uses Clang's APIs to
obtain the required information about the code-base.

Profiling shows that calls to Clang are the major bottleneck in QDoc at
the current time.

To reduce the time spent by Clang, QDoc is being moved to parallelize
and batch certain operations that are related to Clang calls.

`ClangCodeParser`, the component responsible for all of the calls to
Clang libraries, is currently architectured in a way that is not usable
for those optimization purposes.

Hence, it is slowly being modified to be better suited for this
purposes.
As part of those modification, we expect `ClangCodeParser` to be split
into multiple components.

Currently, `ClangCodeParser` is the entry point to a series of
operations that happen in different phases and have very different
restriction on their scopes and their dependencies.

Some of those operation shares some internal state that
an instance of `ClangCodeParser` keeps.

To allow for splitting those various operations up, we need to be able
to pass this shared data around the code-base to all relevant parties.

One such datum is `ClangCodeParser::m_defines`, which is used at various
times to instrument the "-D" arguments for a compilation performed by
Clang's APIs.

Due to the current structure of the code, this processing is performed a
single time by `ClangCodeParser` itself, stored and used multiple times.

The processing steps for `m_defines` that were performed in the
constructor for `ClangCodeParser` is now moved to the main loop of a
QDoc execution.

The constructor for `ClangCodeParser` was modified to obtain a reference
to the required defines data and store it in `m_defines`.

Task-number: QTBUG-111686
Change-Id: I45c2e208090ce29dc249add73e19c96c048f20b3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>