QDoc: Move getDefaultArgs outside ClangCodeParser
When QDoc processes a project, it parses its code-base to obtain and
sanitize the user-provided documentation for the project.

For projects that makes use of C++ code, QDoc uses Clang's APIs to
obtain the required information about the code-base.

Profiling shows that calls to Clang are the major bottleneck in QDoc at
the current time.

To reduce the time spent by Clang, QDoc is being moved to parallelize
and batch certain operations that are related to Clang calls.

`ClangCodeParser`, the component responsible for all of the calls to
Clang libraries, is currently architectured in a way that is not usable
for those optimization purposes.

Hence, it is slowly being modified to be better suited for those
purposes.
As part of those modification, we expect `ClangCodeParser` to be split
into multiple components.

Currently, `ClangCodeParser` is the entry point to a series of
operations that happen in different phases and have very different
restriction on their scopes and their dependencies.

Some of those operation shares some internal state that
an instance of `ClangCodeParser` keeps.

To allow for splitting those various operations up, we need to be able
to pass this shared data around the code-base to all relevant parties.

`ClangCodeParser::getDefaultArgs` populates and internal vector,
`m_args`, one such shared state, to contain the general arguments we
pass to all Clang's based compilations.

As this will be required by multiple parts that are expected to be split
around, `getMoreArgs` was now moved outside `ClangCodeParser` as a free
function.

The dependent state for the function was modified to be passed as
arguments instead of being taken directly from an instance state.

`getDefaultArgs` is called many times in the current QDoc, albeit this
will be lowered in the future, so that it is a very hot path.

`m_args`, that was populated by `getDefaultArgs`, was an instance state
to avoid allocating a new vector many times. As this seems to still be
impactful at the current moment, `getDefaultArgs` now takes a mutable
out-argument that represent the vector where the arguments should be
stored, so that the lifetime of the vector can be managed outside of it,
as `m_args` does.

Usages of `getDefaultArgs` were modified to abide to the new interface.

Task-number: QTBUG-111686
Change-Id: I9da5c8e09606b5b3d665dbed2816e08724012ff1
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>