QDoc: Move getMoreArgs outside ClangCodeParser
When QDoc processes a project, it parses its code-base to obtain and
sanitize the user-provided documentation for the project.

For projects that makes use of C++ code, QDoc uses Clang's APIs to
obtain the required information about the code-base.

Profiling shows that calls to Clang are the major bottleneck in QDoc at
the current time.

To reduce the time spent by Clang, QDoc is being moved to parallelize
and batch certain operations that are related to Clang calls.

`ClangCodeParser`, the component responsible for all of the calls to
Clang libraries, is currently architectured in a way that is not usable
for those optimization purposes.

Hence, it is slowly being modified to be better suited for those
purposes.
As part of those modification, we expect `ClangCodeParser` to be split
into multiple components.

Currently, `ClangCodeParser` is the entry point to a series of
operations that happen in different phases and have very different
restriction on their scopes and their dependencies.

Some of those operation shares some internal state that
an instance of `ClangCodeParser` keeps.

To allow for splitting those various operations up, we need to be able
to pass this shared data around the code-base to all relevant parties.

`ClangCodeParser::getMoreArgs` populates an internal vector,
`m_moreArgs`, one such shared state, to contain some specific arguments we
pass to some Clang's based compilations.

As this will be required by multiple parts that are expected to be split
around, `getMoreArgs` was now moved outside `ClangCodeParser` as a free
function.

The dependent state for the function was modified to be passed as
arguments instead of being taken directly from an instance state.

`getMoreArgs` is called many times in the current QDoc, albeit this
will be lowered in the future, so that it is a very hot path.

`m_moreArgs`, that was populated by `getMoreArgs`, was an instance state
to avoid allocating a new vector many times. As this seems to still be
impactful at the current moment, `getMoreArgs` now takes a mutable
out-argument that represent the vector where the arguments should be
stored, so that the lifetime of the vector can be managed outside of it,
as `m_moreArgs` does.

Usages of `getMoreArgs` were modified to abide to the new interface.

`ClangCodeParser` has a second `m_args` vector that is populated with
default arguments that we use for all compilations.
Every time `m_moreArgs` was populated, its contents would then be
appended to `m_args`, which was used to pass the actual arguments when
compiling.

With the new interface `m_moreArgs` is unnecessary, as we can pass and
populate `m_args` directory in `getMoreArgs`.

Thus `ClangCodeParser::m_moreArgs` is removed.

Task-number: QTBUG-111686
Change-Id: Ic4eebe803b6f8d6e88358f6a51f71e2d07d54b50
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>