Doc: QDoc: Document the \attribution command
The \attribution command is used internally in the output
generated by qtattributionsscanner to mark documentation pages
as attributions.

Although this command is rarely needed to be explicitly used in the
documentation, it is still useful to document; the
\generatelist annotatedattributions can be used for generating lists
of attribution documentation pages, and those lists can be supplemented
using the \attribution command manually.

Fixes: QTBUG-120323
Change-Id: I7f94f5ce9f6a0c0ade41172e41d54b322f6148bf
Reviewed-by: Paul Wicking <paul.wicking@qt.io>