qdoc: Fix markup for template declarations
HTML generator in QDoc constructs a 'marked-up' synopsis for a
function signature.

The code was using an incomplete regular expression to locate
template declarations, in order to replace them with a HTML-encoded
string.

To fix, remove the regular expression(s), and encode the template
declaration directly in CppCodeMarker::markedUpSynopsis() where
the synopsis for the signature is constructed.

In passing, clean up HtmlGenerator::generateSynopsis() a bit.

Fixes: QTBUG-123034
Change-Id: I45684035d10901b01b53088dfdab25d6a0c72033
Reviewed-by: Paul Wicking <paul.wicking@qt.io>