CMake: Fix target-less call of qt_add_lupdate
A call to qt_add_lupdate like the following would lead to an error when
updating translations, because we generated an invalid JSON file and fed
it to lupdate.

    qt_add_lupdate(
        TS_FILES du_de.ts
        SOURCES main.cpp one.cpp two.cpp
    )

Fix this by ensuring that we don't generate subprojects entries in the
JSON file if there aren't any subprojects (== targets).

Pick-to: 6.7.0
Fixes: QTBUG-123011
Change-Id: Ic6465654375a7bfb7d387a6951f8f30408cae452
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit c12750987f2ca474f0016d553e508164957ba659)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>