qdoc: Remove support for subscripted function arguments
QDoc included code for automatically modifying the user-visible
function arguments using '<name>_<number>' syntax to remove the
underscore and display a subscripted number.

Remove this feature; it's undocumented and not used in current
documentation. It was constrained to lowercase a-z argument
names only, numbers 1-9 only. In some of the logic, also zero
was allowed, making the implementation inconsistent.

Eliminating this feature also avoids regular expressions being
constructed and run on every documented function signature and
\a command argument.

Fixes: QTBUG-123072
Change-Id: I9974c124c140f14bb43fa3173defbde083352eae
Reviewed-by: Paul Wicking <paul.wicking@qt.io>