qdoc: Collect \fn parsing failures and process them separately
When deciding whether to output a warning when Clang fails to tie an
\fn signature to a declaration, information about how the base class
is documented is required. Specifically, for classes that are marked
\internal and therefore do not generate output in a standard build,
QDoc omits subsequent warnings for that class's member functions.

This information is no longer available when parsing an \fn topic,
as 'internalness' is meta-information that gets assigned to entities
in a later step.

Introduce a new structure to collect information about parsing
failures, and have ClangCodeParser::parseFnArg() return a variant;
either a matched Node pointer or an error (FnMatchError). This
error contains sufficient context to process it and generate a
warning later on.

Store these errors that occur during topic command processing into a
vector, and handle them after the nodes have also completed
meta-command processing.

The new mechanism can be utilized for consolidating the handling of
multiple types of parsing failures into a central location; for now,
start with match failures for \fn.

For match failures, improve the logic of resolving the parent of a
function from a string representing the function signature; allow
both classes and namespaces, and check the entire parent chain for
the \internal command. Instead of ad-hoc string parsing, use a
regular expression for resolving the scope of an \fn signature.

Simplify the phrasing in the warning message related to \fn matching
failures.

Fixes: QTBUG-123084
Change-Id: I3478774e9141275ad8a043f643850096a146fa1c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>