QDoc: Adapt clang/AST/QualTypeNames.h to upstream change
QDoc maintains a slightly modified version of the Clang header
`QualTypeNames.h`. Due to an upstream change [0] that changes an enum
into an enum class and renames a value that's used in the
aforementioned header, QDoc fails to compile against Clang 18.

This patch introduces a preprocessor conditional check for the major
version of `libclang` and defines the old symbol name if Clang is of a
version where the symbol has been renamed.

[0] - https://github.com/llvm/llvm-project/commit/4ad2ada5216ee2bb3c334a3233a9ab51f2521b82

Fixes: QTBUG-123109
Change-Id: I0fa916a201d963c5f23d87d73105b23fe55719ea
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 8c5e3f46e2fa5056280220c3b0a74025098e2be3)