CMake: Fix discrepancy in qt_add_translations .qm output paths
When calling qt_add_translations with automatic target collection we
1. Predict .qm file paths to be in CMAKE_CURRENT_BINARY_DIR and return
   the paths in QM_FILES_OUTPUT_VARIABLE.
2. Defer the call of qt_add_translations to the end of the PROJECT_DIR
   scope.
3. When running the deferred call, generate .qm files in
   CMAKE_CURRENT_BINARY_DIR.

The values of CMAKE_CURRENT_BINARY_DIR in steps 1 and 3 are different if
the call to qt_add_translations is in a subdirectory.

Fix this by passing the .qm output directory to the deferred call of
qt_add_translations and qt_add_lrelease.

Pick-to: 6.7 6.7.0
Fixes: QTBUG-123405
Change-Id: If4a9cdc03ed4659f28ff70c7662f1125a1f59b55
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>