QDoc: Move ignoresince test
With the advent of QDoc's `tst_validateQdocOutputFiles` test, we now
have a better way to test QDoc's generated output. One of the test
cases in the new test represents a rather comprehensive documentation
project. This project was adopted from `tst_generatedOutput`, where it
is known as `TestCPP`.

This change removes a configuration file from `tst_generatedOutput`
and adapts the aforementioned documentation project such that the
behavior remains tested. Expected output files are modified
accordingly.

Task-number: QTBUG-119500
Pick-to: 6.7
Change-Id: Ifa3c7c4fcbfd05c9ed45f989286fabe79995ad4b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>