QDoc: Unify handling of Cpp and QDoc files
When QDoc parses a project it parses its code-base to extract and
sanitize the user-provided documentation.

Due to recent changes that aim to simplify some upcoming optimization
changes, the way that QDoc handles certain source file lost some
abstraction leading to intermediate duplicated code.

Now that some further changes around this handling were introduced we
can reduce the deduplication and prepare the code-base for further
changes.

A new component, `SourceFileParser` was introduced, that takes care of
dispatching a source file that should be parsed to the component that is
able to parse it.

The code in the main loop that takes care of parsing source files was
modified to make use of the new component.
Some related code was modified in the main loop to better support the
new code and the new flow of data.

Task-number: QTBUG-111686
Change-Id: I5d4a15c62f6563d29abf883a7f71c720040d4577
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>