QDoc: Remove CodeParser from ClangCodeParser
`CodeParser` is a legacy inheritance-based interface for components that
needs to parse a source file.

Recent changes to the code-base moved some of those components to a new
interface so that `CodeParser` is not required anymore.

Hence stop `ClangCodeParser` from inheriting from `CodeParser`.

Some unused methods that were inherited from `CodeParser` and are unused
were removed.

`ClangCodeParser` makes use of a `QDocDatabase` that was previously
obtained from its parent class.
The dependencies is now made explicit and handled as part of the
constructor.

Task-number: QTBUG-111686
Change-Id: Ie4174175f669703cc1aa3fc954efe6f9a1314367
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>