QDoc: Remove CodeParser from PureDocParser
`CodeParser` is a legacy inheritance-based interface for components that
needs to parse a source file.

Recent changes to the code-base moved some of those components to a new
interface so that `CodeParser` is not required anymore.

Hence stop `PureDocParser` from inheriting from `CodeParser`.

Some unused methods that were inherited from `CodeParser` and are unused
were removed.

Task-number: QTBUG-111686
Change-Id: Ie2bf7b0d761750193fdb6d4d13558c6354b4939c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>