qdoc: HTML generator: Use translate="no" only for links to API entities
In most cases, the HTML generator has access to the Node that's
associated with the target when creating a link. If the Node
represents an API entity, add the translate="no" attribute to the
link; otherwise allow translations of the link text.

This improves the translation results as links to section and
page titles are translated as expected.

Fixes: QTBUG-123782
Change-Id: Ic6e0ee5a93f7005a892798d687f1c390f339a724
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>