QDoc: Rename method in Atom
The method `Atom::appendString` operates on a member string list.
However, it doesn't append a string to the list; it concatenates its
argument with the first string in the list. Rename the method to
reduce ambiguity and pave the way for a method that actually appends a
string to the list of strings.

Task-number: QTBUG-122261
Change-Id: I68fa9401e2180c49121f1d2c9853ca149c98cd76
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>