QDoc: Extract method from Tree::resolveTargets (1/4)
`Tree::resolveTargets` is responsible for ensuring the maps
`m_nodesByTargetRef` and `m_nodesByTargetTitle` are correct. The
implementation details for the various tasks clutter reading the
algorithm.

Extract the code that's responsible for ensuring that each unique
PageNode is stored in the `m_pageNodesByTitle` under its title and
delegate to the new method from the call site. As part of the
extraction, replace a manual loop with a call to the std::any_of
algorithm.

Task-number: QTBUG-122261
Change-Id: I391da5bc38d64e6a737a0c77874c9683cce50f48
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>