QDoc: Extract method from Tree::resolveTargets (2/4)
`Tree::resolveTargets` is responsible for ensuring the maps
`m_nodesByTargetRef` and `m_nodesByTargetTitle` are correct. The
implementation details for the various tasks clutter reading the
algorithm.

Extract the code that is responsible for populating the maps for
each section in the table of contents of the Node being processed.

Task-number: QTBUG-122261
Change-Id: I72b9698b2b848bd58eade95981aae2144b536fdc
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>