QDoc: Generate unique references to TOC entries
When QDoc generates the output for a page, the page may contain
multiple `\sectionN` commands. These commands take a string as input,
which becomes the title for the following section of text. If a
section title is duplicated, they must have a unique `id` HTML
attribute to be correctly linkable entities.

This is achieved by extending `Tree::populateTocSectionTargetMap`.
This method iterates over the Atoms that represent the table of
contents for a given Doc for a Node. Ensure that the TOC consists of
uniquely identifiable entries in case of name collisions, such that
QDoc can generate correct output for all such references in a given
document. For section titles that collide on a given page, unique
section references are generated for all but the first instance.
These unique references are then appended to the
Atom::SectionHeadingLeft that follows the Atom being processed, for
later retrieval.

Tree::refForAtom is modified such that, for both Atom::SectionLeft
and Atom::SectionHeadingLeft, the behavior is ensured to be the same.
This is made possible by the fact that these Atoms appear in tandem.
For Atoms of type Atom::SectionHeadingLeft, return the second entry
of its stringlist if its length is two, as that indicates that a
unique reference has been appended to the Atom being processed.

Modify the generator subclasses to account for the new behavior, by
calling the modified Tree::refForAtom to obtain the correct reference
for the Atom being generated.

Add E2E test case for unique anchors for duplicate section titles.

Fixes: QTBUG-122261
Change-Id: I53b4871c31224210b5e9a7cbf3c5df7dec6df8a2
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>