qdoc: validatedQDocOutput autotest: Stop relying on standard library
Two test cases used #include statements from the standard library.
However, the test setup does not ensure that all required include
paths are actually forwarded to QDoc and on to Clang.

Some versions of Clang are able to correctly guess the include
path(s) on some platforms/configurations, but we cannot rely on
that always being the case.

Replace the std lib includes with dummy declarations that should
produce identical results in the expected output.

Change-Id: I6f6b1bb9b176f7fa59ccdfc64c11e1586409eb5a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>