CMake: Fix path of embedded .qm files in sub-directories
Consider a project with sub-directories and a qt_add_translations call
in a sub-directory. Before Qt 6.7.0, the path to the embedded .qm files
was ":/i18n/foo_de.qm". With Qt 6.7.0, the path is
":/i18n/subdir/foo_de.qm".

The reason is that we're called qt6_add_resources with the BASE argument
"${CMAKE_CURRENT_BINARY_DIR}", and that is different if targets are
automatically collected at the end of the project's directory scope.

Fix this by passing the target's BINARY_DIR as BASE instead.

Pick-to: 6.7 6.7.1
Fixes: QTBUG-124188
Change-Id: I6afd39c072e7dccc2ff5937c5cd312a119b68587
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>