QDoc: Warn about multiple '\instantiates' command for same C++ class
The documentation for a QML type, e.g. ```Foo```, can contain the
'\instantiates' command, which takes the name of a C++ class as its
argument:

    \instantiates QFoo

This command establishes a relationship between the nodes that
represent the QML type and C++ class internally to the given QDoc
execution. This relationship is described in the synopsis generated in
the documentation for both the QML type (e.g. "Instantiates: QFoo")
and the C++ class (e.g. "Instantiated by: Foo". In QDoc's node
hierarchy, this relationship exists as a `ClassNode *classNode` for
`QmlTypeNode`s, and as a `QmlTypeNode *qmlElement` for
`ClassNode`s.

The design allows for a unilateral many-to-one relationship, in that
multiple `QmlTypeNode`s can reference the same `ClassNode`, but the
`ClassNode` can only refer to one QML element. As a consequence of
this design choice, when QDoc encounters the '\instantiates' command
with a given C++ class as argument more than once, it replaces the QML
element relationship for the C++ class. This happens without a warning
or other kind of feedback to the user, and may therefore cause
unexpected and surprising output.

QDoc's current implementation does not lend itself to easily change
this into a one-to-many or many-to-many relationship without increased
technical risk. This change introduces a new warning to QDoc in the
instances where this occurs, and extends the QDoc warning
documentation accordingly. The behavior is otherwise kept unchanged,
as QDoc cannot determine the order in a deterministic manner.

Task-number: QTBUG-125076
Change-Id: I5bfc46099f0c4a2f3bf221209c03046859f48a60
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>