qdoc: Introduce the \tm command for handling trademarks
A \tm command takes a string argument and appends
a trademark symbol to it when generating output.

This is done on first-use basis; subsequent uses of \tm
with the same argument will not append the symbol, unless
it's the first occurrence on the page that's being generated.
This way, if marking each occurrence in the source with \tm,
the output will look sensible no matter how/in which order
the source material for the page was compiled.

Implement \tm as a new formatting command. This way, nesting
\tm with other types of formatting (e.g. \e, \c) is automatically
disallowed.

Add a static variable (a set of trademarks) to the Generator base
class in order to track their per-page usage. This set is cleared
whenever initializing a new page.

In HTMLGenerator, link the trademark symbol to a place in the
documentation defined with the `navigation.trademarkspage` configuration
variable. If set, this page acts as the central location for
trademark-related documentation.

[ChangeLog][QDoc] A new command, \tm, was introduced to manage
trademarks in the documentation.

Fixes: QTBUG-124393
Change-Id: I84f54d9696f7f5efd4e82b8b1a99b80a9ff693cf
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>