QDoc: Introduce '\nativetype' command
QDoc supports the use of a command, `\instantiates`, which is meant to
be used to describe what a QML type is called in C++, and generate
links between the C++ documentation and QML documentation. This
command causes some confusion for users, both authors and
readers, as its name describes a concept in programming that goes
beyond "this type is called [Foo] in C++|QML". The output from QDoc is
equally responsible for this confusion.

To avoid causing a breaking change, a new command `\nativetype`, is
introduced as a means to replace the `\instantiates` command, while
allowing for a transition period before the latters removal.

The new command works mostly as the old one, with the following
notable differences:

- Terminology in output is simplified to be as exact as possible. With
  new command, QDoc will generate "In C++"/"In QML" as part of the
  requisites table for C++ class and QML type documentation.
- The new command correctly accounts for the fact that multiple QML
  types can have the same C++ type as their native type. In such
  cases, any one QML type will only list one C++ class under "In C++",
  while the C++ class will contain a list of all corresponding QML
  types.

This change includes an extension of the QDoc output validation test
which exercises the new behavior, as well as documentation
that (hopefully) helps make it clear what this command does.

[ChangeLog][QDoc][General] The new 'nativetype' command supercedes the
'instantiates' command. New documentation should use the new command.

Fixes: QTBUG-125368
Change-Id: I7891d1a9ff8dca3742b60d7bdce32e4d830808a5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>